+++ frontend/src/components/room/CodeEditor.tsx
@@ -18,7 +18,7 @@ function CodeEditor(props: { socket: Socket; room: string }) {
 
       const startPos = target.selectionStart || 0;
       const endPos = target.selectionEnd || 0;
-      const tab = '    ';
+      const tab = '    '; // set to 4 spaces
 
       target.value =
         value.substring(0, startPos) + tab + value.substring(endPos);
