+++ frontend/src/components/room/chat/ChatModel.d.ts
@@ -1,3 +1,11 @@
+export interface ChatModel {
+  text: string;
+  name: string;
+  id: number;
+  socketId: string;
+  room: string;
+}
+
 export interface Roles {
   interviewer: string;
   interviewee: string;
