--- frontend/src/models/ChatModel.tsx
@@ -1,7 +1,0 @@
-export interface ChatModel {
-  text: string;
-  name: string;
-  id: number;
-  socketId: string;
-  room: string;
-}
